Introduce WillPaginate::ArrayMixin
Since changing a whole class (such as Array) globally is often an
anti-pattern, this commit offers to change an object instead,
e.g. `array.extend WillPaginate::ArrayMixin`
